feat: 注释

+++ Entity/src/main/java/indi/midreamsheep/schatapp/backend/service/service/chat/AbstractChatEntity.java
@@ -2,9 +2,7 @@ package indi.midreamsheep.schatapp.backend.service.service.chat;
 
 import lombok.Data;
 
-import java.math.BigInteger;
-
 @Data
 public class AbstractChatEntity {
-    private BigInteger id;
+    private long id;
 }
